Add base BlockUseEvent
