+++ paper-server/patches/sources/net/minecraft/server/level/ServerPlayerGameMode.java.patch
@@ -421,11 +421,11 @@
 +                } finally {
 +                    boolean usedItem = interactionResult1 == null;
 +
-+                    io.papermc.paper.event.block.BlockUseResult result = switch (!usedItem ? interactionResult1 : interactionResult) {
-+                        case net.minecraft.world.InteractionResult.Success ignored -> io.papermc.paper.event.block.BlockUseResult.SUCCESS;
-+                        case net.minecraft.world.InteractionResult.Pass ignored -> io.papermc.paper.event.block.BlockUseResult.PASS;
-+                        case net.minecraft.world.InteractionResult.Fail ignored -> io.papermc.paper.event.block.BlockUseResult.FAIL;
-+                        case net.minecraft.world.InteractionResult.TryEmptyHandInteraction ignored -> io.papermc.paper.event.block.BlockUseResult.TRY_WITH_EMPTY_HAND;
++                    io.papermc.paper.event.block.BlockUseEvent.Result result = switch (!usedItem ? interactionResult1 : interactionResult) {
++                        case InteractionResult.Success ignored -> io.papermc.paper.event.block.BlockUseEvent.Result.SUCCESS;
++                        case InteractionResult.Pass ignored -> io.papermc.paper.event.block.BlockUseEvent.Result.PASS;
++                        case InteractionResult.Fail ignored -> io.papermc.paper.event.block.BlockUseEvent.Result.FAIL;
++                        case InteractionResult.TryEmptyHandInteraction ignored -> io.papermc.paper.event.block.BlockUseEvent.Result.TRY_WITH_EMPTY_HAND;
 +                    };
 +
 +                    new io.papermc.paper.event.block.BlockUsedEvent(
